Move steps_modal components to common/steps_modal.
