+++ src/components/general_layout.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import { themeBreakPoints } from '../util/theme';
 
-import { StepsModalContainer } from './common/steps_modal';
+import { StepsModalContainer } from './common/steps_modal/steps_modal';
 import { Toolbar } from './common/toolbar';
 
 const General = styled.div`
